% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfplot.R
\name{pdfplot}
\alias{pdfplot}
\title{Density plot of the two-parameter exponential distribution}
\usage{
pdfplot(x, theta, beta)
}
\arguments{
\item{x}{vector of quantile.}

\item{theta}{location parameter, where \eqn{\theta > 0}.}

\item{beta}{scale parameter, where \eqn{\beta > 0}.}
}
\value{
a density plot of the two-parameter exponential distribution
}
\description{
Density plot of the two-parameter exponential distribution with \code{theta} and \code{beta}
}
\examples{
x <- seq(0,20,by=0.01)
theta <- 6
beta <- 2
pdfplot(x,theta,beta)
}
