% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplot.R
\name{surplot}
\alias{surplot}
\title{Survival function plot of the two-parameter exponential distribution}
\usage{
surplot(x, theta, beta)
}
\arguments{
\item{x}{vector of quantile.}

\item{theta}{location parameter, where \eqn{\theta > 0}.}

\item{beta}{scale parameter, where \eqn{\beta > 0}.}
}
\value{
a survival function plot of the two-parameter exponential distribution
}
\description{
Survival function plot of the two-parameter exponential distribution with \code{theta} and \code{beta}
}
\examples{
x <- seq(0,20,by=0.01)
theta <- 8
beta <- 1
surplot(x,theta,beta)
}
